% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_codeage.R
\name{age_factor}
\alias{age_factor}
\title{Days to group age.}
\usage{
age_factor(adays)
}
\arguments{
\item{adays}{dbl: days of age.}
}
\value{
ordered factor with age groups.
}
\description{
Days to group age.
}
