% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_weight.R, R/gestational_age.R
\name{birth_weight}
\alias{birth_weight}
\title{Function that classifies birth weight.}
\usage{
birth_weight(gestweeks)

birth_weight(gestweeks)
}
\arguments{
\item{gestweeks}{int: gestational weeks.}

\item{weight}{int: birth weight in gr.}
}
\value{
factor.

factor.
}
\description{
Function that classifies birth weight.

Function that classifies gestational weeks.
}
