% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cie10.R
\name{check_cie10}
\alias{check_cie10}
\title{Function to check cie10 code by}
\usage{
check_cie10(db, age, code_age, code_cie10, sex, ...)
}
\arguments{
\item{db}{dataset. death.}

\item{age}{age.}

\item{sex}{sex.}

\item{...}{vars for identification.}

\item{code_age.}{1 years. 2 months. 3 days. 4 hours.}

\item{code_cie10.}{Vars with code of 3 or 4 letters.}
}
\description{
Function to check cie10 code by
}
