% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_check.R
\name{cie_check}
\alias{cie_check}
\title{Function to check CIE code by range of age, Suspected maternal death,
Valid cause of death, unlikely cause of death and sex restriction.}
\usage{
cie_check(db, age, code_age, code_cie10, sex, ...)
}
\arguments{
\item{db}{dataset. death.}

\item{age}{age.}

\item{sex}{sex.}

\item{...}{vars for identification.}

\item{code_age.}{1 years. 2 months. 3 days. 4 hours.}

\item{code_cie10.}{Vars with code of 3 or 4 letters.}
}
\description{
Function to check CIE code by range of age, Suspected maternal death,
Valid cause of death, unlikely cause of death and sex restriction.
}
