% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cie10.R
\name{cie_summary.cie_check}
\alias{cie_summary.cie_check}
\title{Create summary report on console}
\usage{
\method{cie_summary}{cie_check}(x)
}
\arguments{
\item{x}{object class cie_check}
}
\description{
Create summary report on console
}
