% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_check.R
\name{cie_tbl_enos.cie_check}
\alias{cie_tbl_enos.cie_check}
\title{create table with Notifiable infectous diseases.}
\usage{
\method{cie_tbl_enos}{cie_check}(x)
}
\arguments{
\item{x}{object class cie_check}
}
\value{
tibble.
}
\description{
create table with Notifiable infectous diseases.
}
