% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cie10.R
\name{cie_tbl_errors.cie_check}
\alias{cie_tbl_errors.cie_check}
\title{create table with errors}
\usage{
\method{cie_tbl_errors}{cie_check}(x)
}
\arguments{
\item{x}{object class cie_check}
}
\value{
tibble.
}
\description{
create table with errors
}
