% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_redu.R
\name{code_redu}
\alias{code_redu}
\title{This is a helper function. This function builds the code_redu
by joining the variables year and group age to later join with the tbl_critred.}
\usage{
code_redu(year, grupedad, codmuer)
}
\arguments{
\item{year}{Year}

\item{grupedad}{Group age with modes M1(0-6 days), M2(7-28 days) and M3(29d-11m).}

\item{codmuer}{death code.}
}
\value{
code_redu for join with tbl_critred.
}
\description{
This is a helper function. This function builds the code_redu
by joining the variables year and group age to later join with the tbl_critred.
}
