% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_useless.R
\name{code_useless}
\alias{code_useless}
\title{Function that create variable with Type of useless code from CIE10.}
\usage{
code_useless(x)
}
\arguments{
\item{x}{chr cie10 four digits code.}
}
\description{
Function that create variable with Type of useless code from CIE10.
}
