% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{col_length}
\alias{col_length}
\title{Return a message if all columns are or not are present by compare intersection length.}
\usage{
col_length(tbl_1, vital_fact = "deaths")
}
\arguments{
\item{tbl_1}{tbl for check.}

\item{vital_fact}{vital fact of dataset. Options: 'deaths' or 'births'. Default 'deaths.}
}
\description{
Return a message if all columns are or not are present by compare intersection length.
}
