% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness_tbl.R
\name{completeness_tbl}
\alias{completeness_tbl}
\title{Check completeness from a vector of known cats.}
\usage{
completeness_tbl(name, var, cats)
}
\arguments{
\item{name:}{chr name of var.}

\item{var:}{vector with data.}

\item{cats:}{vector with categories}
}
\description{
Check completeness from a vector of known cats.
}
