% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestational_age.R
\name{gestational_age}
\alias{gestational_age}
\title{Function that classifies gestational weeks.}
\usage{
gestational_age(gestweeks)
}
\arguments{
\item{gestweeks}{int: gestational weeks.}
}
\value{
factor.
}
\description{
Function that classifies gestational weeks.
}
