% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_codeage.R
\name{rec_age2day}
\alias{rec_age2day}
\title{This is a helper function that transform age and age code to days.}
\usage{
rec_age2day(age, code_age)
}
\arguments{
\item{age}{age. Integer.}

\item{code_age}{code age Assume code_age levels 1 = years, 2 = months,
3 = days, 4 = hours, 5 = minutes.}
}
\description{
This is a helper function that transform age and age code to days.
}
