% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_redu.R
\name{redu_summary.data.frame}
\alias{redu_summary.data.frame}
\title{join with the tbl_critred2.}
\usage{
\method{redu_summary}{data.frame}(t)
}
\arguments{
\item{tblReductible}{tbl result from reductible.}
}
\description{
join with the tbl_critred2.
}
