% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_redu.R
\name{reductible}
\alias{reductible}
\title{join with the tbl_critred2.}
\usage{
reductible(df, age, codeage, year, codmuer, ...)
}
\arguments{
\item{age}{age.}

\item{codeage}{Group age with modes M1(0-6 days), M2(7-28 days) and M3(29d-11m).}

\item{year}{Year}

\item{codmuer}{death code.}
}
\value{
tibble.
}
\description{
join with the tbl_critred2.
}
