% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerofill.R
\name{zerofill}
\alias{zerofill}
\title{Fill with left zeros}
\usage{
zerofill(int, nceros)
}
\arguments{
\item{int}{integer}

\item{n}{int length of character}
}
\value{
character of length \code{n} filled with 0 on the left
}
\description{
Fill with left zeros
}
