% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_codeage.R
\name{age_codeage}
\alias{age_codeage}
\title{recode age group from two variables age and code_age. Assume code_age levels 1 = years, 2 = months, 3 = days, 4 = hours, 5 = minutes.}
\usage{
age_codeage(age, code_age)
}
\arguments{
\item{age}{num: numeric age.}

\item{code_age}{num: numeric code_age}
}
\value{
ordered factor with age groups.
}
\description{
recode age group from two variables age and code_age. Assume code_age levels 1 = years, 2 = months, 3 = days, 4 = hours, 5 = minutes.
}
