% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_weight.R
\name{birth_weight}
\alias{birth_weight}
\title{Function that classifies birth weight.}
\usage{
birth_weight(weight)
}
\arguments{
\item{weight}{int: birth weight in gr.}
}
\value{
factor.
}
\description{
Function that classifies birth weight.
}
